PYTHON/drf/MyProject/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
Django==4.0.4
django-extensions==3.1.5
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
